Remove Javassist leftovers
