+++ stagemonitor-core/src/main/java/org/stagemonitor/core/CorePlugin.java
@@ -31,7 +31,6 @@ import org.stagemonitor.core.configuration.converter.SetValueConverter;
 import org.stagemonitor.core.elasticsearch.ElasticsearchClient;
 import org.stagemonitor.core.elasticsearch.IndexSelector;
 import org.stagemonitor.core.grafana.GrafanaClient;
-import org.stagemonitor.core.instrument.MainStagemonitorClassFileTransformer;
 import org.stagemonitor.core.metrics.MetricNameFilter;
 import org.stagemonitor.core.metrics.MetricsAggregationReporter;
 import org.stagemonitor.core.metrics.MetricsWithCountFilter;
@@ -335,7 +334,7 @@ public class CorePlugin extends StagemonitorPlugin {
 			.key("stagemonitor.instrument.excludedInstrumenter")
 			.dynamic(false)
 			.label("Excluded Instrumenters")
-			.description("A list of the simple class names of StagemonitorJavassistInstrumenters that should not be applied")
+			.description("A list of the simple class names of StagemonitorByteBuddyTransformers that should not be applied")
 			.defaultValue(Collections.<String>emptySet())
 			.configurationCategory(CORE_PLUGIN_NAME)
 			.build();
@@ -393,7 +392,6 @@ public class CorePlugin extends StagemonitorPlugin {
 
 	@Override
 	public void initializePlugin(InitArguments initArguments) {
-		MainStagemonitorClassFileTransformer.clearClassPools();
 		this.metricRegistry = initArguments.getMetricRegistry();
 		final Integer reloadInterval = getReloadConfigurationInterval();
 		if (reloadInterval > 0) {
