+++ stagemonitor-core/src/main/java/org/stagemonitor/core/Stagemonitor.java
@@ -15,7 +15,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.stagemonitor.core.configuration.Configuration;
 import org.stagemonitor.core.configuration.source.ConfigurationSource;
-import org.stagemonitor.core.instrument.MainStagemonitorClassFileTransformer;
+import org.stagemonitor.core.instrument.AgentAttacher;
 import org.stagemonitor.core.metrics.metrics2.Metric2Filter;
 import org.stagemonitor.core.metrics.metrics2.Metric2Registry;
 import org.stagemonitor.core.util.ClassUtils;
@@ -243,7 +243,7 @@ public final class Stagemonitor {
 			reloadPluginsAndConfiguration();
 		}
 		tryStartMonitoring();
-		onShutdownActions.add(MainStagemonitorClassFileTransformer.performRuntimeAttachment());
+		onShutdownActions.add(AgentAttacher.performRuntimeAttachment());
 	}
 
 	private static void tryStartMonitoring() {
