+++ stagemonitor-core/src/main/java/org/stagemonitor/core/metrics/annotations/MeteredTransformer.java
@@ -11,14 +11,11 @@ import java.util.Collections;
 import java.util.List;
 
 import com.codahale.metrics.annotation.Metered;
-import javassist.CtClass;
-import javassist.CtMethod;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 import org.stagemonitor.core.Stagemonitor;
 import org.stagemonitor.core.instrument.StagemonitorByteBuddyTransformer;
-import org.stagemonitor.core.metrics.aspects.SignatureUtils;
 
 /**
  * Implementation for the {@link Metered} annotation
@@ -28,18 +25,6 @@ public class MeteredTransformer extends StagemonitorByteBuddyTransformer {
 	@Override
 	protected ElementMatcher.Junction<? super MethodDescription.InDefinedShape> getExtraMethodElementMatcher() {
 		return isAnnotatedWith(Metered.class);
-	}
-
-	public void transformClass(CtClass ctClass, ClassLoader loader) throws Exception {
-		for (CtMethod ctMethod : ctClass.getDeclaredMethods()) {
-			Metered Metered = (Metered) ctMethod.getAnnotation(Metered.class);
-			if (Metered != null) {
-				final String signature = SignatureUtils.getSignature(ctMethod, Metered.name(),
-						Metered.absolute());
-				ctMethod.insertBefore("org.stagemonitor.core.metrics.annotations.MeteredInstrumenter" +
-						".meter(\"" + signature + "\");");
-			}
-		}
 	}
 
 	@Advice.OnMethodEnter
